refactoring and reducing amount of ios web platforms tested
