+++ webdriverio/web/webdriver/test/configs/wdio.saucelabs.ios.conf.js
@@ -1,43 +1,26 @@
-const {config} = require('./wdio.shared.conf');
-const build = `iOS Safari WebdriverIO-V6 build-${new Date().getTime()}`
-const appiumVersion = process.env.npm_config_appium_version
-
-// =========================
-// Sauce Simulator specific config
-// =========================
-config.user = process.env.SAUCE_USERNAME;
-config.key = process.env.SAUCE_ACCESS_KEY;
-// If you run your tests on Sauce Labs you can specify the region you want to run your tests
-// in via the `region` property. Available short handles for regions are `us` (default) and `eu`.
-// These regions are used for the Sauce Labs VM cloud and the Sauce Labs Real Device Cloud.
-// If you don't provide the region, it defaults to `us`.
-config.region = process.env.REGION || 'us';
+const {config} = require('./wdio.shared.sauce.mobile.conf');
+const build = `iOS Safari WebdriverIO-V6 Appium ${config.appiumVersion} build-${new Date().getTime()}`;
 
 // ============
 // Capabilities
 // ============
+// For all capabilities please check
+// http://appium.io/docs/en/writing-running-appium/caps/#general-capabilities
+//
+// For configuring an Emulator please check
+// https://wiki.saucelabs.com/display/DOCS/Platform+Configurator#/
 config.capabilities = [
     /**
      * iOS Safari
      */
     {
-        name: 'iOS 13.0 iPhone safari',
         build: build,
         browserName: 'safari',
         platformName: 'iOS',
-        platformVersion: '13.0',
+        platformVersion: '12.0',
         deviceName: 'iPhone Simulator',
     },
     {
-        name: 'iOS 13.2 iPhone safari',
-        build: build,
-        browserName: 'safari',
-        platformName: 'iOS',
-        platformVersion: '13.2',
-        deviceName: 'iPhone Simulator',
-    },
-    {
-        name: 'iOS 13.4 iPhone safari',
         build: build,
         browserName: 'safari',
         platformName: 'iOS',
@@ -45,51 +28,17 @@ config.capabilities = [
         deviceName: 'iPhone Simulator',
     },
     {
-        name: 'iOS 14.0 iPhone safari',
         build: build,
         browserName: 'safari',
         platformName: 'iOS',
         platformVersion: '14.0',
         deviceName: 'iPhone Simulator',
     },
-    {
-        name: 'iOS 13.0 iPad safari',
-        build: build,
-        browserName: 'safari',
-        platformName: 'iOS',
-        platformVersion: '13.0',
-        deviceName: 'iPad Simulator',
-    },
-    {
-        name: 'iOS 13.2 iPad safari',
-        build: build,
-        browserName: 'safari',
-        platformName: 'iOS',
-        platformVersion: '13.2',
-        deviceName: 'iPad Simulator',
-    },
-    {
-        name: 'iOS 13.4 iPad safari',
-        build: build,
-        browserName: 'safari',
-        platformName: 'iOS',
-        platformVersion: '13.4',
-        deviceName: 'iPad Simulator',
-    },
-    {
-        name: 'iOS 14.0 iPad safari',
-        build: build,
-        browserName: 'safari',
-        platformName: 'iOS',
-        platformVersion: '14.0',
-        deviceName: 'iPad Simulator',
-    }
 ];
 
-if (appiumVersion !== undefined) {
+if (config.appiumVersion !== undefined && config.appiumVersion !== 'default') {
     config.capabilities.forEach(capability => {
-        capability.name += ` - Appium Version ${appiumVersion}`
-        capability.appiumVersion = appiumVersion
+        capability.appiumVersion = config.appiumVersion;
     });
 }
